Implement subject management and link categorization
- Update `Subject` and `PaperLink` entities to use `Long` for IDs and fix `PaperLinkFts` column mapping for `displayName`.
- Expand `PaperLinkDao` and `PaperLinkRepository` with full CRUD operations for subjects and reactive data streams.
- Update `NavGraph` to include `SubjectsDestination` and register the new `SubjectsScreen`.
- Add a `TopAppBar` to `HomeScreen` with a navigation action to the subjects management screen.
- Refactor `AddLinkScreen` and `AddLinkViewModel` to support subject selection and on-the-fly subject creation via a nested dialog.
- Externalize hardcoded strings to `strings.xml` and update icons to `AutoMirrored` versions for better RTL support.
- Improve `AddLinkViewModel` logic to persist URI permissions and automatically select newly created subjects.

+++ app/src/main/java/com/joasasso/paperlink/data/local/PaperLink.kt
@@ -35,5 +35,5 @@ data class PaperLink(
     @ColumnInfo(name = "display_name") val displayName: String? = null,
     val note: String? = null,
     @ColumnInfo(name = "created_at") val createdAt: Long = System.currentTimeMillis(),
-    @ColumnInfo(name = "subject_id") val subjectId: Int? = null
+    @ColumnInfo(name = "subject_id") val subjectId: Long? = null
 )
